Refactor 674.
