--- leetcode/java/src/L674_LongestContinuousIncreasingSubsequence.java
@@ -1,27 +1,0 @@
-/**
- * Given an unsorted array of integers, find the length of longest continuous increasing
- * subsequence.
- */
-public class L674_LongestContinuousIncreasingSubsequence {
-
-    /**
-     * Find the length of longest continuous increasing subsequence.
-     * @param nums Array to be checked.
-     * @return
-     */
-    public int findLengthOfLCIS(int[] nums) {
-        int i = 0, j = 1;
-        int max = 0;
-        int tmp;
-        while (i < nums.length) {
-            while (j < nums.length && nums[j] > nums[j - 1])
-                j++;
-            tmp = j - i;
-            if (tmp > max)
-                max = tmp;
-            i = j;
-            j += 1;
-        }
-        return max;
-    }
-}
